reorganized all the files into an easier way to merge things
