--- vehicle/mkviii/software/lv/bspd/BUILD
@@ -1,4 +1,0 @@
-
-exports_files([
-    "bspd.yml",
-])
