--- vehicle/mkviii/software/lv/throttle/throttle.h
@@ -1,90 +1,0 @@
-#ifndef THROTTLE_H
-#define THROTTLE_H
-
-#include "stm32g4xx_hal.h"
-
-#define PA1_GPIO_Port   GPIOA
-#define PA1_Pin         GPIO_PIN_1
-
-
-void HAL_FDCAN_MspInit(FDCAN_HandleTypeDef *hfdcan) {
-    if (hfdcan->Instance == FDCAN1) {
-        // Enable clocks for FDCAN1 and GPIOA
-        __HAL_RCC_FDCAN_CLK_ENABLE();
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-
-        // Configure PA11 (RX) and PA12 (TX) for Alternate Function 9 (FDCAN1)
-        GPIO_InitTypeDef GPIO_InitStruct = {0};
-        GPIO_InitStruct.Pin = GPIO_PIN_11 | GPIO_PIN_12;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-        GPIO_InitStruct.Alternate = GPIO_AF9_FDCAN1;
-        HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-    }
-}
-
-
-// Function Prototypes
-static void GpioInit(void) {
-  __HAL_RCC_GPIOF_CLK_ENABLE();
-  __HAL_RCC_GPIOA_CLK_ENABLE();
-
-  GPIO_InitTypeDef GPIO_InitStruct = {0};
-
-  // Uses PA1_... from throttle.h
-  HAL_GPIO_WritePin(PA1_GPIO_Port, PA1_Pin, GPIO_PIN_RESET);
-
-  GPIO_InitStruct.Pin = PA1_Pin;
-  GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-  GPIO_InitStruct.Pull = GPIO_NOPULL;
-  GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-  HAL_GPIO_Init(PA1_GPIO_Port, &GPIO_InitStruct);
-}
-
-
-void ErrorHandler(void) { // this will only happen if error on board
-  __disable_irq();
-  HAL_GPIO_WritePin(PA1_GPIO_Port, PA1_Pin, GPIO_PIN_SET);
-  while (1) {
-  }
-}
-
-void SystemClockConfig(void) { // if you are messing with this please consult someone
-  RCC_OscInitTypeDef RCC_OscInitStruct = {0};
-  RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
-
-  HAL_PWREx_ControlVoltageScaling(PWR_REGULATOR_VOLTAGE_SCALE1);
-
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSE;
-  RCC_OscInitStruct.HSEState = RCC_HSE_ON;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSE;
-  RCC_OscInitStruct.PLL.PLLM = RCC_PLLM_DIV1;
-  RCC_OscInitStruct.PLL.PLLN = 8;
-  RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV2;
-  RCC_OscInitStruct.PLL.PLLQ = RCC_PLLQ_DIV2;
-  RCC_OscInitStruct.PLL.PLLR = RCC_PLLR_DIV2;
-
-  if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK) {
-    ErrorHandler();
-  }
-
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK|RCC_CLOCKTYPE_SYSCLK
-                              |RCC_CLOCKTYPE_PCLK1|RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV1;
-  RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-
-  if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_2) != HAL_OK) {
-    ErrorHandler();
-  }
-}
-
-void SysTick_Handler(void) {
-  HAL_IncTick();
-}
-
-
-#endif // THROTTLE_H
