--- vehicle/mkviii/software/sensing/sus_ex/BUILD
@@ -1,5 +1,0 @@
-
-
-exports_files([
-    "sus_ex.yml",
-])
