--- vehicle/mkviii/software/sensing/sus_ex/sus_ex.yml
@@ -1,17 +1,0 @@
-name: sus_strain
-subscribe:
-  - name: throttle
-    mob: 1
-publish:
-  - name: sus_strain_l
-    id: 0x490
-    freq_hz: 100
-    signals:
-      - name: sus_strain
-        slice: 0 + 64
-        unit:
-          type: uint64_t
-          name: 'ticks'
-          offset: 0
-          scale: 1
-
